Changed the internals of convert fn slightly
Added doc strings to all the vars and functions.
Added all the missing units to css-unit-regex.
Cleaned up some random comments and fixme's.
Updated the to input of functions to properly reflect what they are, that being to-unit.
Updated (and improved?) the function specs for convert and unit.
Added unit-keywords to the css specs namespace.